requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.4
pydantic==2.10.4
python-dateutil==2.9.0.post0
feedparser==6.0.11
requests==2.32.3